requirements.txt
Flask
Flask-SQLAlchemy
gunicorn
SQLAlchemy
psycopg2-binary
python-dotenv
requests